Agrego metodo para hacer mas performante obtener el resultado para exportar el yml. Cambio Integer por double al atributo de tiempo de uso en resultado. Agrego un test para saber que el promedio en la nueva manera de calcularlo  de igual que antes.

+++ src/untref/aydoo/procesadorestadistico/Resultado.java
@@ -7,7 +7,7 @@ public class Resultado {
 	private double tiempoPromedioDeUso;
 	private Recorrido recorridoMasVecesRealizado;
 	private Bicicleta bicicletaUtilizadaMasTiempo;
-	private Integer tiempoUsoBicicletaMasUsada;
+	private double tiempoUsoBicicletaMasUsada;
 
 	public Resultado(Bicicleta bicicletaUtilizadaMasVeces,
 			Bicicleta bicicletaUtilizadaMenosVeces, double tiempoPromedioDeUso,
@@ -118,11 +118,11 @@ public class Resultado {
 		return bicicletaUtilizadaMasTiempo;
 	}
 
-	public void setTiempoUsoBicicletaMasUsada(Integer tiempoUsoBicicletaMasUsada) {
+	public void setTiempoUsoBicicletaMasUsada(Double tiempoUsoBicicletaMasUsada) {
 		this.tiempoUsoBicicletaMasUsada = tiempoUsoBicicletaMasUsada;
 	}
 
-	public Integer getTiempoUsoBicicletaMasUsada() {
+	public Double getTiempoUsoBicicletaMasUsada() {
 		return tiempoUsoBicicletaMasUsada;
 	}
